strlen int to size_t

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: raqroca- <raqroca-@student.42.madrid.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/13 12:53:26 by raqroca-          #+#    #+#             */
-/*   Updated: 2026/01/19 12:03:01 by raqroca-         ###   ########.fr       */
+/*   Updated: 2026/01/23 13:26:59 by raqroca-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
